src/namenode/requirements.txt
grpcio-tools==1.47.0
Flask==2.1.2
